rustwide: extract crater::logs into rustwide::logging
