+++ Cargo.toml
@@ -66,9 +66,13 @@ hmac = "0.7"
 sha-1 = "0.8"
 rust_team_data = { git = "https://github.com/rust-lang/team" }
 systemstat = "0.1.4"
+rustwide = { path = "rustwide" }
 
 [dev-dependencies]
 assert_cmd = "0.10.1"
 difference = "2.0.0"
 num_cpus = "1.8.0"
 predicates = "1.0.0"
+
+[workspace]
+members = ["rustwide"]
