+++ rustwide/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "rustwide"
+version = "0.1.0"
+authors = ["Pietro Albini <pietro@pietroalbini.org>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4.6"
