+++ rustwide/src/lib.rs
@@ -1,0 +1,8 @@
+#![warn(missing_docs)]
+
+//! rustwide is a library to execute your code against the Rust ecosystem, powering projects like
+//! [Crater][crater].
+//!
+//! [crater]: https://github.com/rust-lang/crater
+
+pub mod logging;
