+++ src/results/db.rs
@@ -2,7 +2,6 @@ use crate::config::Config;
 use crate::crates::{Crate, GitHubRepo};
 use crate::db::{Database, QueryUtils};
 use crate::experiments::Experiment;
-use crate::logs::{self, LogStorage};
 use crate::prelude::*;
 use crate::results::{
     DeleteResults, EncodedLog, EncodingType, ReadResults, TestResult, WriteResults,
@@ -10,6 +9,7 @@ use crate::results::{
 use crate::toolchain::Toolchain;
 use base64;
 use log::LevelFilter;
+use rustwide::logging::{self, LogStorage};
 use serde_json;
 use std::collections::HashMap;
 
@@ -208,8 +208,14 @@ impl<'a> WriteResults for DatabaseDB<'a> {
     where
         F: FnOnce() -> Fallible<TestResult>,
     {
-        let storage = existing_logs.unwrap_or_else(|| LogStorage::new(LevelFilter::Info, config));
-        let result = logs::capture(&storage, f)?;
+        let storage = existing_logs.unwrap_or_else(|| {
+            LogStorage::new(
+                LevelFilter::Info,
+                config.sandbox.build_log_max_size.to_bytes(),
+                config.sandbox.build_log_max_lines,
+            )
+        });
+        let result = logging::capture(&storage, f)?;
         let output = storage.to_string();
         self.store_result(
             ex,
@@ -321,7 +327,7 @@ mod tests {
 
     #[test]
     fn test_results() {
-        crate::logs::init_test();
+        rustwide::logging::init();
 
         let db = Database::temp().unwrap();
         let results = DatabaseDB::new(&db);
