+++ src/runner/mod.rs
@@ -10,13 +10,13 @@ use crate::config::Config;
 use crate::crates::Crate;
 use crate::docker::DockerEnv;
 use crate::experiments::{Experiment, Mode};
-use crate::logs::LogStorage;
 use crate::prelude::*;
 use crate::results::{TestResult, WriteResults};
 use crate::runner::graph::build_graph;
 use crate::runner::worker::{DiskSpaceWatcher, Worker};
 use crate::utils;
 use crossbeam_utils::thread::{scope, ScopedJoinHandle};
+use rustwide::logging::LogStorage;
 use std::collections::HashMap;
 use std::path::Path;
 use std::sync::Mutex;
